Remove use of can-stache-bindings
